# src/swagger.json
{
	"components": {
...
		"schemas": {
			"ProfileType": {
				"type": "string",
				"enum": [
					"stakeholder",
					"admin",
					"unit owner"
				]
			},
			"RecordStatus": {
				"type": "string",
				"enum": [
					"active",
					"inactive"
				]
			},
...
			"CommunityAttr": {
				"properties": {
					"description": {
						"type": "string"
					}
				},
				"required": [
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChargeType": {
				"type": "string",
				"enum": [
					"unit",
					"percentage",
					"amount"
				]
			},
			"PostingType": {
				"type": "string",
				"enum": [
					"monthly",
					"manual",
					"accrued",
					"interest",
					"quarterly"
				]
			},
			"ChargeAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"communityId": {
						"type": "number",
						"format": "double"
					},
					"community": {
						"$ref": "#/components/schemas/CommunityAttr"
					},
					"code": {
						"type": "string"
					},
					"rate": {
						"type": "number",
						"format": "double"
					},
					"chargeType": {
						"$ref": "#/components/schemas/ChargeType"
					},
					"postingType": {
						"$ref": "#/components/schemas/PostingType"
					},
					"thresholdInMonths": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"priority": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"passOn": {
						"type": "boolean",
						"nullable": true
					}
				},
				"required": [
					"communityId",
					"code",
					"rate",
					"chargeType",
					"postingType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TransactionType": {
				"type": "string",
				"enum": [
					"charged",
					"collected"
				]
			},
...
			"PaymentType": {
				"type": "string",
				"enum": [
					"cash",
					"check"
				]
			},
			"ProfileAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"mobileNumber": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/ProfileType"
					},
					"status": {
						"$ref": "#/components/schemas/RecordStatus"
					},
					"scopes": {
						"type": "string"
					},
					"remarks": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"username",
					"password",
					"email",
					"type",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DisbursementAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"voucherId": {
						"type": "number",
						"format": "double"
					},
					"purchaseOrderId": {
						"type": "number",
						"format": "double"
					},
					"chargeId": {
						"type": "number",
						"format": "double"
					},
					"releasedBy": {
						"type": "number",
						"format": "double"
					},
					"paymentType": {
						"$ref": "#/components/schemas/PaymentType"
					},
					"details": {
						"type": "string"
					},
					"checkNumber": {
						"type": "string"
					},
					"checkPostingDate": {
						"type": "string",
						"format": "date-time"
					},
					"checkIssuingBank": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"releasedByProfile": {
						"$ref": "#/components/schemas/ProfileAttr"
					},
					"charge": {
						"$ref": "#/components/schemas/ChargeAttr"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"chargeId",
					"releasedBy",
					"paymentType",
					"details",
					"amount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FieldError": {
				"properties": {
					"value": {
						"type": "string",
						"nullable": true
					},
					"field": {
						"type": "string",
						"nullable": true
					},
					"type": {
						"type": "string",
						"nullable": true
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"value",
					"field",
					"type",
					"message"
				],
				"type": "object"
			},
			"EntityError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"fieldErrors": {
						"items": {
							"$ref": "#/components/schemas/FieldError"
						},
						"type": "array",
						"default": []
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PropertyAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "string"
					},
					"floorArea": {
						"type": "number",
						"format": "double"
					},
					"address": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/RecordStatus"
					}
				},
				"required": [
					"code",
					"floorArea",
					"address",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaymentDetailAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"collectedBy": {
						"type": "number",
						"format": "double"
					},
					"orNumber": {
						"type": "string"
					},
					"paymentType": {
						"$ref": "#/components/schemas/PaymentType"
					},
					"checkNumber": {
						"type": "string"
					},
					"checkPostingDate": {
						"type": "string",
						"format": "date-time"
					},
					"checkIssuingBank": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"collectedBy",
					"orNumber",
					"paymentType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TransactionAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"chargeId": {
						"type": "number",
						"format": "double"
					},
					"charge": {
						"$ref": "#/components/schemas/ChargeAttr"
					},
					"propertyId": {
						"type": "number",
						"format": "double"
					},
					"property": {
						"$ref": "#/components/schemas/PropertyAttr"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"transactionPeriod": {
						"type": "string",
						"format": "date-time"
					},
					"transactionType": {
						"$ref": "#/components/schemas/TransactionType"
					},
					"waivedBy": {
						"type": "number",
						"format": "double"
					},
					"comments": {
						"type": "string"
					},
					"paymentDetailId": {
						"type": "number",
						"format": "double"
					},
					"paymentDetail": {
						"$ref": "#/components/schemas/PaymentDetailAttr"
					},
					"rateSnapshot": {
						"type": "number",
						"format": "double"
					},
					"batchId": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"chargeId",
					"propertyId",
					"amount",
					"transactionPeriod",
					"transactionType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Month": {
				"type": "string",
				"enum": [
					"JAN",
					"FEB",
					"MAR",
					"APR",
					"MAY",
					"JUN",
					"JUL",
					"AUG",
					"SEP",
					"OCT",
					"NOV",
					"DEC"
				]
			},
...
			"RequestStatus": {
				"type": "string",
				"enum": [
					"approved",
					"rejected",
					"pending",
					"cancelled"
				]
			},
			"ExpenseAttr": {
				"properties": {
					"voucherId": {
						"type": "number",
						"format": "double"
					},
					"purchaseRequestId": {
						"type": "number",
						"format": "double"
					},
					"purchaseOrderId": {
						"type": "number",
						"format": "double"
					},
					"categoryId": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"unitCost": {
						"type": "number",
						"format": "double"
					},
					"totalCost": {
						"type": "number",
						"format": "double"
					},
					"waivedBy": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"categoryId",
					"category",
					"description",
					"quantity",
					"unitCost",
					"totalCost"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreatePurchaseOrder": {
				"properties": {
					"purchaseRequestId": {
						"type": "number",
						"format": "double"
					},
					"otherDetails": {
						"type": "string"
					},
					"fulfillmentDate": {
						"type": "string",
						"format": "date-time"
					},
					"vendorName": {
						"type": "string"
					}
				},
				"required": [
					"purchaseRequestId",
					"otherDetails",
					"fulfillmentDate",
					"vendorName"
				],
				"type": "object"
			},
			"CreateVoucherOrOrder": {
				"properties": {
					"orderData": {
						"$ref": "#/components/schemas/CreatePurchaseOrder"
					},
					"expenses": {
						"items": {
							"$ref": "#/components/schemas/ExpenseAttr"
						},
						"type": "array"
					},
					"chargeId": {
						"type": "number",
						"format": "double"
					},
					"requestedDate": {
						"type": "string",
						"format": "date-time"
					},
					"requestedBy": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"expenses",
					"chargeId",
					"requestedDate",
					"requestedBy",
					"description"
				],
				"type": "object"
			},
			"ApproveVoucherOrOrder": {
				"properties": {
					"disbursements": {
						"items": {
							"$ref": "#/components/schemas/DisbursementAttr"
						},
						"type": "array"
					},
					"codes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"purchaseOrderId": {
						"type": "number",
						"format": "double"
					},
					"purchaseRequestId": {
						"type": "number",
						"format": "double"
					},
					"voucherId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"codes"
				],
				"type": "object"
			},
...
			"PurchaseRequestAttr": {
				"properties": {
					"series": {
						"type": "string"
					},
					"approverProfiles": {
						"items": {
							"$ref": "#/components/schemas/ProfileAttr"
						},
						"type": "array"
					},
					"rejectedByProfile": {
						"$ref": "#/components/schemas/ProfileAttr"
					},
					"requestedByProfile": {
						"$ref": "#/components/schemas/ProfileAttr"
					},
					"expenses": {
						"items": {
							"$ref": "#/components/schemas/ExpenseAttr"
						},
						"type": "array"
					},
					"comments": {
						"type": "string",
						"nullable": true
					},
					"rejectedBy": {
						"type": "number",
						"format": "double"
					},
					"approvedBy": {
						"type": "string",
						"nullable": true
					},
					"requestedDate": {
						"type": "string",
						"format": "date-time"
					},
					"requestedBy": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/RequestStatus"
					},
					"totalCost": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"chargeId": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"series",
					"requestedDate",
					"requestedBy",
					"status",
					"totalCost",
					"description",
					"chargeId"
				],
				"type": "object"
			},
			"RejectPurchaseRequest": {
				"properties": {
					"rejectedBy": {
						"type": "number",
						"format": "double"
					},
					"comments": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"rejectedBy",
					"comments",
					"id"
				],
				"type": "object"
			},
			"SettingAttr": {
				"properties": {
					"key": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"key",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoryAttr": {
				"properties": {
					"subCategories": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"communityId": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"subCategories",
					"description",
					"communityId"
				],
				"type": "object"
			},
			"PostTransactionBody": {
				"properties": {
					"batchId": {
						"type": "string"
					},
					"propertyId": {
						"type": "number",
						"format": "double"
					},
					"month": {
						"$ref": "#/components/schemas/Month"
					},
					"year": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"propertyId",
					"month",
					"year"
				],
				"type": "object"
			},
			"PostCollectionBody": {
				"properties": {
					"transactions": {
						"items": {
							"$ref": "#/components/schemas/TransactionAttr"
						},
						"type": "array"
					},
					"paymentDetail": {
						"$ref": "#/components/schemas/PaymentDetailAttr"
					}
				},
				"required": [
					"transactions",
					"paymentDetail"
				],
				"type": "object"
			},
			"Period": {
				"properties": {
					"month": {
						"$ref": "#/components/schemas/Month"
					},
					"year": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"month",
					"year"
				],
				"type": "object"
			},
			"PropertyCollectionByChargeView": {
				"properties": {
					"chargeCode": {
						"type": "string"
					},
					"collected": {
						"type": "number",
						"format": "double"
					},
					"floorArea": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"chargeCode",
					"collected",
					"floorArea",
					"address",
					"code",
					"id"
				],
				"type": "object"
			},
			"RefundPaymentBody": {
				"properties": {
					"comments": {
						"type": "string"
					},
					"refundedBy": {
						"type": "number",
						"format": "double"
					},
					"paymentDetailId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"comments",
					"refundedBy",
					"paymentDetailId"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"bearer": {
				"type": "apiKey",
				"in": "header",
				"name": "authorization"
			}
		}
	},
	"info": {
		"title": "luminous-api",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/purchase-request/getPurchaseRequest/{id}": {
			"get": {
				"operationId": "GetPurchaseRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PurchaseRequestAttr"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/purchase-request/getAllPurchaseRequestsByChargeAndStatus/{chargeId}/{status}": {
			"get": {
				"operationId": "GetAllPurchaseRequestsByChargeAndStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PurchaseRequestAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "chargeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "status",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RequestStatus"
						}
					}
				]
			}
		},
		"/api/purchase-request/postPurchaseRequest": {
			"post": {
				"operationId": "PostPurchaseRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateVoucherOrOrder"
							}
						}
					}
				}
			}
		},
		"/api/purchase-request/updatePurchaseRequest/{id}": {
			"patch": {
				"operationId": "UpdatePurchaseRequest",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateVoucherOrOrder"
							}
						}
					}
				}
			}
		},
		"/api/purchase-request/approvePurchaseRequest": {
			"post": {
				"operationId": "ApprovePurchaseRequest",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ApproveVoucherOrOrder"
							}
						}
					}
				}
			}
		},
		"/api/purchase-request/rejectPurchaseRequest": {
			"post": {
				"operationId": "RejectPurchaseRequest",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RejectPurchaseRequest"
							}
						}
					}
				}
			}
		},
		"/api/purchase-request/notifyPurchaseRequestApprovers/{id}": {
			"post": {
				"operationId": "NotifyPurchaseRequestApprovers",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/api/setting/getSettingValue": {
			"get": {
				"operationId": "GetSettingValue",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/setting/updateSettingValue": {
			"patch": {
				"operationId": "UpdateSettingValue",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SettingAttr"
							}
						}
					}
				}
			}
		},
		"/api/setting/getAllCategories": {
			"get": {
				"operationId": "GetAllCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CategoryAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": []
			}
		},
		"/api/setting/updateCategories": {
			"patch": {
				"operationId": "UpdateCategories",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/CategoryAttr"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/api/transaction/postMonthlyCharges": {
			"post": {
				"operationId": "PostMonthlyCharges",
				"responses": {
					"201": {
						"description": "created"
					},
					"400": {
						"description": "duplicate charges",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostTransactionBody"
							}
						}
					}
				}
			}
		},
		"/api/transaction/postCollections": {
			"post": {
				"operationId": "PostCollections",
				"responses": {
					"201": {
						"description": "created"
					},
					"400": {
						"description": "bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityError"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostCollectionBody"
							}
						}
					}
				}
			}
		},
		"/api/transaction/postTransactions": {
			"post": {
				"operationId": "PostTransactions",
				"responses": {
					"201": {
						"description": "created"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/TransactionAttr"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/api/transaction/getAvailablePeriods/{propertyId}": {
			"get": {
				"operationId": "GetAvailablePeriods",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Period"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "propertyId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/transaction/suggestPaymentBreakdown/{propertyId}": {
			"get": {
				"operationId": "SuggestPaymentBreakdown",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TransactionAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "propertyId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "amount",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "year",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "month",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Month"
						}
					}
				]
			}
		},
		"/api/transaction/getCollectionBreakdown/{year}/{month}": {
			"get": {
				"operationId": "GetCollectionBreakdown",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PropertyCollectionByChargeView"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "year",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "month",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Month"
						}
					}
				]
			}
		},
		"/api/transaction/refundPayment/{propertyId}": {
			"patch": {
				"operationId": "RefundPayment",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "propertyId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RefundPaymentBody"
							}
						}
					}
				}
			}
		},
		"/api/transaction/getWaterReadingByPeriod/{year}/{month}": {
			"get": {
				"operationId": "GetWaterReadingByPeriod",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TransactionAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "year",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "month",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Month"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}